+++ src/Renderer/Fragment/Browser/BrowserFragmentAsset/detect-input.ts
