+++ src/Renderer/Fragment/Issues/IssuesFragment.tsx
@@ -2,18 +2,18 @@ import electron from 'electron';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import moment from 'moment';
-import {SystemStreamEvent} from '../Event/SystemStreamEvent';
-import {StreamEvent} from '../Event/StreamEvent';
-import {LibraryStreamEvent} from '../Event/LibraryStreamEvent';
-import {IssueRepo} from '../Repository/IssueRepo';
-import {IssueEvent} from '../Event/IssueEvent';
-import {SystemStreamRepo} from '../Repository/SystemStreamRepo';
-import {StreamRepo} from '../Repository/StreamRepo';
-import {WebViewEvent} from '../Event/WebViewEvent';
-import {FilterHistoryRepo} from '../Repository/FilterHistoryRepo';
-import {Color} from '../../Util/Color';
-import {GARepo} from '../Repository/GARepo';
-import {ConfigRepo} from '../Repository/ConfigRepo';
+import {SystemStreamEvent} from '../../Event/SystemStreamEvent';
+import {StreamEvent} from '../../Event/StreamEvent';
+import {LibraryStreamEvent} from '../../Event/LibraryStreamEvent';
+import {IssueRepo} from '../../Repository/IssueRepo';
+import {IssueEvent} from '../../Event/IssueEvent';
+import {SystemStreamRepo} from '../../Repository/SystemStreamRepo';
+import {StreamRepo} from '../../Repository/StreamRepo';
+import {WebViewEvent} from '../../Event/WebViewEvent';
+import {FilterHistoryRepo} from '../../Repository/FilterHistoryRepo';
+import {Color} from '../../../Util/Color';
+import {GARepo} from '../../Repository/GARepo';
+import {ConfigRepo} from '../../Repository/ConfigRepo';
 
 const remote = electron.remote;
