+++ src/Renderer/Fragment/Stream/LibraryStreamsFragment.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import electron from 'electron';
-import {SystemStreamEvent} from '../Event/SystemStreamEvent';
-import {StreamEvent} from '../Event/StreamEvent';
-import {IssueEvent} from '../Event/IssueEvent';
-import {LibraryStreamRepo} from '../Repository/LibraryStreamRepo';
-import {LibraryStreamEvent} from '../Event/LibraryStreamEvent';
-import {IssueRepo} from '../Repository/IssueRepo';
-import {GARepo} from '../Repository/GARepo';
+import {SystemStreamEvent} from '../../Event/SystemStreamEvent';
+import {StreamEvent} from '../../Event/StreamEvent';
+import {IssueEvent} from '../../Event/IssueEvent';
+import {LibraryStreamRepo} from '../../Repository/LibraryStreamRepo';
+import {LibraryStreamEvent} from '../../Event/LibraryStreamEvent';
+import {IssueRepo} from '../../Repository/IssueRepo';
+import {GARepo} from '../../Repository/GARepo';
 
 const remote = electron.remote;
 const MenuItem = remote.MenuItem;
