+++ src/Renderer/Fragment/Stream/ModalStreamSettingFragment.tsx
@@ -1,31 +1,37 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import {StreamEvent} from '../Event/StreamEvent';
-import {StreamRepo} from '../Repository/StreamRepo';
-import {GARepo} from '../Repository/GARepo';
+import electron from 'electron';
+import {StreamEvent} from '../../Event/StreamEvent';
+import {StreamRepo} from '../../Repository/StreamRepo';
+import {GARepo} from '../../Repository/GARepo';
+import {ConfigRepo} from '../../Repository/ConfigRepo';
 
 interface State {
   queries: string[];
 }
 
-export class FilteredStreamSettingFragment extends React.Component<any, State> {
+export class ModalStreamSettingFragment extends React.Component<any, State> {
   state: State = {queries: []};
-  private readonly _streamListenerIds: number[] = [];
   private _stream: any = null;
-  private _filteredStream: any = null;
+  private readonly _streamListenerIds: number[] = [];
   private _originalHeight: string = null;
 
   componentDidMount() {
     {
       let id;
-      id = StreamEvent.addOpenFilteredStreamSettingListener(this._show.bind(this));
+      id = StreamEvent.addOpenStreamSettingListener(this._show.bind(this));
       this._streamListenerIds.push(id);
     }
 
+    electron.ipcRenderer.on('create-new-stream', (_ev, stream)=>{
+      this._show(stream, true);
+    });
+
     const dialog = ReactDOM.findDOMNode(this);
     this._originalHeight = window.getComputedStyle(dialog).height;
+
     dialog.addEventListener('close', ()=>{
-      StreamEvent.emitCloseFilteredStreamSetting(this._stream);
+      StreamEvent.emitCloseStreamSetting(this._stream);
     });
   }
 
@@ -33,29 +39,28 @@ export class FilteredStreamSettingFragment extends React.Component<any, State> {
     StreamEvent.removeListeners(this._streamListenerIds);
   }
 
-  _show(stream, filter, filteredStream) {
+  _show(stream, asNewStream = false) {
     this._stream = stream;
-    this._filteredStream = filteredStream;
     const dialog = ReactDOM.findDOMNode(this);
-    if (filteredStream) {
-      dialog.querySelector('#nameInput').value = filteredStream.name;
-      dialog.querySelector('#filterInput').value = filteredStream.filter;
-      dialog.querySelector('#notificationInput').checked = filteredStream.notification === 1;
-      dialog.querySelector('#colorInput').value = filteredStream.color;
-      dialog.querySelector('.icon-flow-cascade').style.color = filteredStream.color;
-    } else {
-      dialog.querySelector('#nameInput').value = `- My Filter`;
-      dialog.querySelector('#filterInput').value = filter;
+    let queries;
+    if (stream) {
+      queries = JSON.parse(stream.queries);
+      dialog.querySelector('#nameInput').value = stream.name;
       dialog.querySelector('#notificationInput').checked = stream.notification === 1;
       dialog.querySelector('#colorInput').value = stream.color;
-      dialog.querySelector('.icon-flow-cascade').style.color = stream.color;
+      dialog.querySelector('.icon-github').style.color = stream.color;
+    } else {
+      queries = [''];
+      dialog.querySelector('#nameInput').value = '';
+      dialog.querySelector('#notificationInput').checked = true;
+      dialog.querySelector('#colorInput').value = '';
+      dialog.querySelector('.icon-github').style.color = null;
     }
 
-    dialog.querySelector('#streamName').textContent = stream.name;
-    const queries = JSON.parse(stream.queries);
-    this._updateHeight(queries.length);
-    this.setState({queries});
+    if (asNewStream) this._stream = null;
 
+    this.setState({queries});
+    this._updateHeight(queries.length);
     dialog.showModal();
   }
 
@@ -66,37 +71,74 @@ export class FilteredStreamSettingFragment extends React.Component<any, State> {
   }
 
   _handleCancel() {
+    this.setState({queries: []});
     const dialog = ReactDOM.findDOMNode(this);
     dialog.close();
   }
 
   async _handleOK() {
     const name = ReactDOM.findDOMNode(this).querySelector('#nameInput').value;
-    const filter = ReactDOM.findDOMNode(this).querySelector('#filterInput').value;
     const notification = ReactDOM.findDOMNode(this).querySelector('#notificationInput').checked ? 1 : 0;
     const color = ReactDOM.findDOMNode(this).querySelector('#colorInput').value;
+
+    // pick up queries from each DOMs
+    const queries = [];
+    {
+      let index = 0;
+      while (1) {
+        const el = ReactDOM.findDOMNode(this).querySelector(`#queryInput${index}`);
+        if (el && el.value) {
+          queries.push(el.value);
+        } else {
+          break;
+        }
+        index++;
+      }
+    }
 
     if (color && !color.match(/^#[0-9A-Fa-f]{3,6}$/)) {
       return;
     }
 
-    if (name && filter) {
+    if (name && queries.length) {
+      this.setState({queries: []});
       const dialog = ReactDOM.findDOMNode(this);
       dialog.close();
 
-      if (this._filteredStream) {
-        StreamRepo.rewriteFilteredStream(this._filteredStream.id, name, filter, notification, color);
+      if (this._stream) {
+        StreamRepo.rewriteStream(this._stream.id, name, queries, notification, color);
       } else {
-        await StreamRepo.createFilteredStream(this._stream, name, filter, notification, color);
-        GARepo.eventFilteredStreamCreate();
+        await StreamRepo.createStream(name, queries, notification, color);
+        GARepo.eventStreamCreate(queries.length);
       }
     }
+  }
+
+  _handleHelp() {
+    electron.shell.openExternal('https://jasperapp.io/doc.html#stream');
+  }
+
+  _handlePreview() {
+    const query = ReactDOM.findDOMNode(this).querySelector('#queryInput0').value;
+    if (!query) return;
+
+    const apiHost = ConfigRepo.getConfig().github.host;
+    let webHost = null;
+    if (apiHost === 'api.github.com') {
+      webHost = 'github.com';
+    } else {
+      webHost = apiHost;
+    }
+
+    const url = `https://${webHost}/search?s=updated&o=desc&type=Issues&q=${encodeURIComponent(query)}`;
+    const proxy = window.open(url, 'github-search-preview', 'width=1024px,height=600px');
+    proxy.focus();
   }
 
   _handleColor() {
     // hack: dom operation
     const color = ReactDOM.findDOMNode(this).querySelector('#colorInput').value;
-    const icon = ReactDOM.findDOMNode(this).querySelector('.icon-flow-cascade');
+    const icon = ReactDOM.findDOMNode(this).querySelector('.icon-github');
     icon.style.color = color;
   }
 
@@ -107,35 +149,38 @@ export class FilteredStreamSettingFragment extends React.Component<any, State> {
     this._handleColor();
   }
 
-  _handleHelp() {
-    const shell = require('electron').shell;
-    shell.openExternal('https://jasperapp.io/doc.html#filter');
+  _handleAddQuery() {
+    const queries = this.state.queries;
+    queries.push('');
+    this.setState({queries});
+    this._updateHeight(queries.length);
   }
 
   render() {
     const queryNodes = this.state.queries.map((query, index) => {
-      return <div key={index}>{query}</div>;
+      return <input key={index} id={`queryInput${index}`} className="form-control"
+                    defaultValue={query}
+                    placeholder="is:pr author:octocat"/>;
     });
 
     return (
-      <dialog className="stream-setting filtered-stream-setting">
+      <dialog className="stream-setting">
         <div className="window">
           <div className="window-content">
 
             <div>
-              <div className="form-group from-stream">
-                <label><span>Stream: </span><span id="streamName"/></label>
+              <div className="form-group" title="stream name">
+                <label>Name</label>
+                <input id="nameInput" className="form-control" placeholder="stream name"/>
+              </div>
+
+              <div className="form-group queries" title="stream query">
+                <div className="queries-section">
+                  <label>Query <span className="help-link" onClick={this._handleHelp.bind(this)}>help</span></label>
+                  <span className="flex-stretch"/>
+                  <span className="icon icon-plus" onClick={this._handleAddQuery.bind(this)}/>
+                </div>
                 {queryNodes}
-              </div>
-
-              <div className="form-group" title="filter name">
-                <label>Filter Name</label>
-                <input id="nameInput" className="form-control" placeholder="filtered stream name"/>
-              </div>
-
-              <div className="form-group" title="filter query">
-                <label>Filter <span className="help-link" onClick={this._handleHelp.bind(this)}>help</span></label>
-                <input id="filterInput" className="form-control" placeholder="is:pr author:octocat"/>
               </div>
 
               <div className="form-group" title="stream icon color">
@@ -160,13 +205,17 @@ export class FilteredStreamSettingFragment extends React.Component<any, State> {
                   <span className="color-palette" title="#AC8EF6" style={{background: '#AC8EF6'}} onClick={this._handleColorPalette.bind(this)}/>
                 </span>
                 <input id="colorInput" className="form-control" placeholder="#aabbcc" onKeyUp={this._handleColor.bind(this)}/>
-                <span className="icon icon-flow-cascade"/>
+                <span className="icon icon-github"/>
               </div>
 
-              <div className="form-actions split-buttons">
+              <div className="form-group" title="stream notification">
                 <label>
                   <input type="checkbox" id="notificationInput"/> Notification
                 </label>
+              </div>
+
+              <div className="form-actions split-buttons">
+                <button className="btn btn-form btn-default" onClick={this._handlePreview.bind(this)}>Preview</button>
                 <span className="flex-stretch"/>
                 <button className="btn btn-form btn-default" onClick={this._handleCancel.bind(this)}>Cancel</button>
                 <button className="btn btn-form btn-primary" onClick={this._handleOK.bind(this)}>OK</button>
