+++ src/Renderer/Fragment/Stream/StreamsFragment.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import electron from 'electron';
-import {StreamRepo} from '../Repository/StreamRepo';
-import {LibraryStreamEvent} from '../Event/LibraryStreamEvent';
-import {SystemStreamEvent} from '../Event/SystemStreamEvent';
-import {StreamEvent} from '../Event/StreamEvent';
-import {IssueEvent} from '../Event/IssueEvent';
-import {IssueRepo} from '../Repository/IssueRepo';
-import {GARepo} from '../Repository/GARepo';
+import {StreamRepo} from '../../Repository/StreamRepo';
+import {LibraryStreamEvent} from '../../Event/LibraryStreamEvent';
+import {SystemStreamEvent} from '../../Event/SystemStreamEvent';
+import {StreamEvent} from '../../Event/StreamEvent';
+import {IssueEvent} from '../../Event/IssueEvent';
+import {IssueRepo} from '../../Repository/IssueRepo';
+import {GARepo} from '../../Repository/GARepo';
 
 const remote = electron.remote;
 const MenuItem = remote.MenuItem;
